requirements.txt
pascal-voc-writer
pillow
torch